remove git related code
