+++ services/cd-service/pkg/repository/repository_test.go
@@ -671,216 +671,6 @@ func TestApplyQueue(t *testing.T) {
 	}
 }
 
-func getTransformer(i int) (Transformer, error) {
-	transformerType := i % 5
-	switch transformerType {
-	case 0:
-	case 1:
-	case 2:
-		return &CreateApplicationVersion{
-			Application: "foo",
-			Manifests: map[string]string{
-				"development": fmt.Sprintf("%d", i),
-			},
-			Version: uint64(i + 1),
-		}, nil
-	case 3:
-		return &ErrorTransformer{}, TransformerError
-	case 4:
-		return &InvalidJsonTransformer{}, InvalidJson
-	}
-	return &ErrorTransformer{}, TransformerError
-}
-
-func createGitWithCommit(remote string, local string, t *testing.B) {
-	cmd := exec.Command("git", "init", "--bare", remote)
-	cmd.Start()
-	cmd.Wait()
-
-	cmd = exec.Command("git", "clone", remote, local) // Clone git dir
-	_, err := cmd.Output()
-	if err != nil {
-		t.Fatal(err)
-	}
-	cmd = exec.Command("touch", "a") // Add a new file to git
-	cmd.Dir = local
-	_, err = cmd.Output()
-	if err != nil {
-		t.Fatal(err)
-	}
-	cmd = exec.Command("git", "add", "a") // Add a new file to git
-	cmd.Dir = local
-	_, err = cmd.Output()
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	cmd = exec.Command("git", "commit", "-m", "adding") // commit the new file
-	cmd.Dir = local
-	cmd.Env = []string{
-		"GIT_AUTHOR_NAME=kuberpult",
-		"GIT_COMMITTER_NAME=kuberpult",
-		"EMAIL=test@kuberpult.com",
-	}
-	_, err = cmd.Output()
-	if err != nil {
-		if exitErr, ok := err.(*exec.ExitError); ok {
-			t.Logf("stderr: %s\n", exitErr.Stderr)
-			t.Logf("stderr: %s\n", err)
-		}
-		t.Fatal(err)
-	}
-	cmd = exec.Command("git", "push", "origin", "HEAD") // push the new commit
-	cmd.Dir = local
-	_, err = cmd.Output()
-	if err != nil {
-		t.Fatal(err)
-	}
-}
-
-func BenchmarkApplyQueue(t *testing.B) {
-	t.StopTimer()
-	dir := t.TempDir()
-	remoteDir := path.Join(dir, "remote")
-	localDir := path.Join(dir, "local")
-	createGitWithCommit(remoteDir, localDir, t)
-
-	repo, err := New(
-		testutil.MakeTestContext(),
-		RepositoryConfig{
-			URL:                 "file://" + remoteDir,
-			Path:                localDir,
-			ArgoCdGenerateFiles: true,
-		},
-	)
-	if err != nil {
-		t.Fatalf("new: expected no error, got '%e'", err)
-	}
-	repoInternal := repo.(*repository)
-	// The worker go routine is now blocked. We can move some items into the queue now.
-	results := make([]<-chan error, t.N)
-	expectedResults := make([]error, t.N)
-	expectedReleases := make(map[int]bool, t.N)
-	tf, _ := getTransformer(0)
-	repoInternal.Apply(testutil.MakeTestContext(), tf)
-
-	t.StartTimer()
-	for i := 0; i < t.N; i++ {
-		tf, expectedResult := getTransformer(i)
-		results[i] = repoInternal.applyDeferred(testutil.MakeTestContext(), tf)
-		expectedResults[i] = expectedResult
-		if expectedResult == nil {
-			expectedReleases[i+1] = true
-		}
-	}
-
-	for i := 0; i < t.N; i++ {
-		if err := <-results[i]; err != expectedResults[i] {
-			t.Errorf("result[%d] expected error \"%v\" but got \"%v\"", i, expectedResults[i], err)
-		}
-	}
-	releases, _ := repo.State().Releases("foo")
-	if !cmp.Equal(expectedReleases, convertToSet(releases)) {
-		t.Fatal("Output mismatch (-want +got):\n", cmp.Diff(expectedReleases, convertToSet(releases)))
-	}
-}
-
-func TestPushUpdate(t *testing.T) {
-	tcs := []struct {
-		Name            string
-		InputBranch     string
-		InputRefName    string
-		InputStatus     string
-		ExpectedSuccess bool
-	}{
-		{
-			Name:            "Should succeed",
-			InputBranch:     "main",
-			InputRefName:    "refs/heads/main",
-			InputStatus:     "",
-			ExpectedSuccess: true,
-		},
-		{
-			Name:            "Should fail because wrong branch",
-			InputBranch:     "main",
-			InputRefName:    "refs/heads/master",
-			InputStatus:     "",
-			ExpectedSuccess: false,
-		},
-		{
-			Name:            "Should fail because status not empty",
-			InputBranch:     "master",
-			InputRefName:    "refs/heads/master",
-			InputStatus:     "i am the status, stopping this from working",
-			ExpectedSuccess: false,
-		},
-	}
-	for _, tc := range tcs {
-		tc := tc
-		t.Run(tc.Name, func(t *testing.T) {
-			t.Parallel()
-			var success = false
-			actualError := defaultPushUpdate(tc.InputBranch, &success)(tc.InputRefName, tc.InputStatus)
-			if success != tc.ExpectedSuccess {
-				t.Fatal(fmt.Sprintf("expected sucess=%t but got %t", tc.ExpectedSuccess, success))
-			}
-			if actualError != nil {
-				t.Fatal(fmt.Sprintf("expected no error but got %s but got none", actualError))
-			}
-		})
-	}
-}
-
-// func TestDeleteDirIfEmpty(t *testing.T) {
-// 	tcs := []struct {
-// 		Name           string
-// 		CreateThisDir  string
-// 		DeleteThisDir  string
-// 		ExpectedError  error
-// 		ExpectedReason SuccessReason
-// 	}{
-// 		{
-// 			Name:           "Should succeed: dir exists and is empty",
-// 			CreateThisDir:  "foo/bar",
-// 			DeleteThisDir:  "foo/bar",
-// 			ExpectedReason: NoReason,
-// 		},
-// 		{
-// 			Name:           "Should succeed: dir does not exist",
-// 			CreateThisDir:  "foo/bar",
-// 			DeleteThisDir:  "foo/bar/pow",
-// 			ExpectedReason: DirDoesNotExist,
-// 		},
-// 		{
-// 			Name:           "Should succeed: dir does not exist",
-// 			CreateThisDir:  "foo/bar/pow",
-// 			DeleteThisDir:  "foo/bar",
-// 			ExpectedReason: DirNotEmpty,
-// 		},
-// 	}
-// 	for _, tc := range tcs {
-// 		tc := tc
-// 		t.Run(tc.Name, func(t *testing.T) {
-// 			t.Parallel()
-// 			repo := setupRepositoryTest(t)
-// 			state := repo.State()
-// 			err := state.Filesystem.MkdirAll(tc.CreateThisDir, 0777)
-// 			if err != nil {
-// 				t.Fatalf("error in mkdir: %v", err)
-// 				return
-// 			}
-
-// 			successReason, err := state.DeleteDirIfEmpty(tc.DeleteThisDir)
-// 			if diff := cmp.Diff(tc.ExpectedError, err, cmpopts.EquateErrors()); diff != "" {
-// 				t.Errorf("error mismatch (-want, +got):\n%s", diff)
-// 			}
-// 			if successReason != tc.ExpectedReason {
-// 				t.Fatal("Output mismatch (-want +got):\n", cmp.Diff(tc.ExpectedReason, successReason))
-// 			}
-// 		})
-// 	}
-// }
-
 func TestProcessQueueOnce(t *testing.T) {
 	tcs := []struct {
 		Name           string
