+++ services/cd-service/pkg/repository/transformer.go
@@ -1537,9 +1537,7 @@ func (c *DeleteEnvironmentLock) Transform(
 		return "", err
 	}
 	s := State{
-		Commit:               nil,
 		MinorRegexes:         state.MinorRegexes,
-		Filesystem:           state.Filesystem,
 		MaxNumThreads:        state.MaxNumThreads,
 		DBHandler:            state.DBHandler,
 		ReleaseVersionsLimit: state.ReleaseVersionsLimit,
@@ -2162,7 +2160,6 @@ func (c *DeployApplicationVersion) Transform(
 	if err != nil {
 		return "", err
 	}
-	fs := state.Filesystem
 
 	var manifestContent []byte
 	version, err := state.DBHandler.DBSelectReleaseByVersion(ctx, transaction, c.Application, c.Version, true)
@@ -2303,10 +2300,8 @@ func (c *DeployApplicationVersion) Transform(
 	}
 	t.AddAppEnv(c.Application, c.Environment, teamOwner)
 	s := State{
-		Commit:               nil,
 		MinorRegexes:         state.MinorRegexes,
 		MaxNumThreads:        state.MaxNumThreads,
-		Filesystem:           fs,
 		DBHandler:            state.DBHandler,
 		ReleaseVersionsLimit: state.ReleaseVersionsLimit,
 		CloudRunClient:       state.CloudRunClient,
